kH/s value fixed

+++ src-qt/widget.h
@@ -14,7 +14,7 @@ class Widget : public QWidget
 
 public:
     Widget(QWidget *parent = nullptr);
-    void setLog(std::string tm, uint64_t tt, uint64_t f, double k);
+    void setLog(std::string tm, uint64_t tt, uint64_t f, uint64_t k);
     void setAddr(std::string address);
     option conf;
     ~Widget();
@@ -22,7 +22,7 @@ public:
 private:
     Ui::Widget *ui;
     unsigned int m_mode = 1;
-    double speedRecord = 0;
+    uint64_t speedRecord = 0;
     void restoreNotABugLinkButton();
 
 public slots:
